feat: implement delete devices by room functionality and update related services

+++ OwnLightSystem/OwnLight.AutomationService/AutomationService.Application/Features/Room/Handlers/Commands/DeleteRoomCommandHandler.cs
@@ -1,3 +1,4 @@
+using AutomationService.Application.Common.Services.Interfaces;
 using AutomationService.Application.Features.Room.Commands;
 using AutomationService.Domain.Interfaces;
 using MediatR;
@@ -7,11 +8,13 @@ namespace AutomationService.Application.Features.Room.Handlers.Commands;
 
 public class DeleteRoomCommandHandler(
     IRoomRepository roomRepository,
-    IHttpContextAccessor httpContextAccessor
+    IHttpContextAccessor httpContextAccessor,
+    IDeviceServiceClient deviceServiceClient
 ) : IRequestHandler<DeleteRoomCommand>
 {
     private readonly IRoomRepository _roomRepository = roomRepository;
     private readonly IHttpContextAccessor _httpContextAccessor = httpContextAccessor;
+    private readonly IDeviceServiceClient _deviceServiceClient = deviceServiceClient;
 
     public async Task<Unit> Handle(DeleteRoomCommand request, CancellationToken cancellationToken)
     {
@@ -26,7 +29,22 @@ public class DeleteRoomCommandHandler(
         if (room.UserId.ToString() != userId)
             throw new UnauthorizedAccessException("Esse cômodo não pertence ao usuário logado.");
 
+        var accessToken = _httpContextAccessor
+            .HttpContext?.Request.Headers["Authorization"]
+            .ToString()
+            .Replace("Bearer ", "");
+
+        if (string.IsNullOrEmpty(accessToken))
+            throw new UnauthorizedAccessException("JWT token ausente.");
+
+        var response = await _deviceServiceClient.DeleteDevicesByRoomIdAsync(room.Id, accessToken);
+        if (!response.IsSuccess)
+            throw new Exception(
+                $"Falha ao deletar dispositivos do cômodo na API DeviceService: {response.ErrorMessage}"
+            );
+
         await _roomRepository.DeleteAsync(room.Id, cancellationToken);
+
         return Unit.Value;
     }
 }
